Rename recipeFinder.js to recipefinder.js
