+++ frontend/pages/recipefinder.html
@@ -87,7 +87,7 @@
       </div>
     </main>
 
-    <script type="module" src="../script/recipeFinder.js"></script>
+    <script type="module" src="../script/recipefinder.js"></script>
     <script type="module" src="../script/user.js"></script>
     <script src="https://code.jquery.com/jquery-3.4.1.slim.min.js" integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
